# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PVT OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/role/{role}/role_permissions": {
            "get": {
                "tags": [
                    "ROLES"
                ],
                "summary": "LISTADO DE PERMISOS ASIGNADOS A UN ROL",
                "operationId": "role_permissions",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id permiso",
                        "required": false
                    },
                    {
                        "name": "operation_id",
                        "in": "query",
                        "description": "id de la operacion",
                        "required": false
                    },
                    {
                        "name": "action_id",
                        "in": "query",
                        "description": "id de la acción ",
                        "required": false
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtro por nombre de permiso ",
                        "required": false
                    },
                    {
                        "name": "display_name",
                        "in": "query",
                        "description": "Filtro del nombre de visualización del permiso",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina a mostrar",
                        "required": false,
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Por Pagina",
                        "required": false,
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/user": {
            "get": {
                "tags": [
                    "USUARIO"
                ],
                "summary": "LISTADO DE USUARIOS",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina a mostrar",
                        "required": false
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Por Pagina",
                        "required": false,
                        "example": 10
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Usuarios activos(1) o inactivos(0) ",
                        "required": false,
                        "example": 1
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Filtro por nombre de Usuario ",
                        "required": false
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Filtro por Apellidos",
                        "required": false
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "Filtro por cargo",
                        "required": false
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "filtro por nombre de Usuario",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "USUARIO"
                ],
                "summary": "NUEVO USUARIO",
                "description": "Creación de un nuevo usuario",
                "operationId": "crear usuario",
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "nombre de usuario required",
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "description": "nombres required",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "apellidos required",
                                        "type": "string"
                                    },
                                    "identity_card": {
                                        "description": "carnet de identidad required",
                                        "type": "string"
                                    },
                                    "position": {
                                        "description": "Cargo del usuario required",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "telefono required",
                                        "type": "integer"
                                    },
                                    "city_id": {
                                        "description": "id de ciudad",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/user/{user}/role": {
            "patch": {
                "tags": [
                    "USUARIO"
                ],
                "summary": "ESTABLECER O ELIMINAR EL ROL A UN USUARIO",
                "operationId": "setOrRemoveRolForUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 240
                    }
                ],
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_id": {
                                        "description": "id rol required",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}